Move new logic into a somewhat reasonable place
