+++ crates/autopilot/src/database/ethflow_events/mod.rs
@@ -1,0 +1,4 @@
+//! Contains all the components to handle events that are relevant for the ethflow which are not
+//! already part of the generic CoWSwapOnchainOrders interface.
+pub mod event_retriever;
+pub mod event_storing;
